test 2 - refactor
