+++ src/Infrastructure/VolleyM.Infrastructure.EventBroker/EventHandlerWrapper.cs
@@ -1,25 +1,22 @@
-﻿using System.Reflection;
+﻿using System;
+using System.Reflection;
 using System.Threading.Tasks;
 using VolleyM.Domain.Contracts.Crosscutting;
 
 namespace VolleyM.Infrastructure.EventBroker
 {
-    public class EventHandlerWrapper : IEventHandler<object>
+    public class EventHandlerWrapper
     {
-        //ToDo: Remove this field
-        private readonly object _handler;
         private readonly MethodInfo _handleMethod;
 
-        public EventHandlerWrapper(object handler)
+        public EventHandlerWrapper(Type handlerType)
         {
-            _handler = handler;
-
-            _handleMethod = _handler.GetType().GetMethod(nameof(IEventHandler<object>.Handle));
+            _handleMethod = handlerType.GetMethod(nameof(IEventHandler<object>.Handle));
         }
 
-        public Task Handle(object @event)
+        public Task Handle(object handler, object @event)
         {
-            return (Task)_handleMethod.Invoke(_handler, new[] { @event });
+            return (Task)_handleMethod.Invoke(handler, new[] { @event });
         }
     }
 }
